refactor: move API utility helpers into utils
