+++ frontend/src/api/accounts/requests.ts
@@ -1,5 +1,5 @@
 import { authenticatedFetch } from '@/api/client';
-import { buildQueryString } from '@/api/queryString';
+import { buildQueryString } from '@/api/utils/queryString';
 import type {
   Account,
   AccountBalanceSnapshot,
