+++ frontend/src/api/merchants/requests.ts
@@ -1,5 +1,5 @@
 import { authenticatedFetch } from '@/api/client';
-import { buildQueryString, type QueryStringValue } from '@/api/queryString';
+import { buildQueryString, type QueryStringValue } from '@/api/utils/queryString';
 import type {
   CreateMerchantPayload,
   Merchant,
