+++ frontend/src/api/taxAdvantagedCategories/hooks.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { runWithMinimumPendingTime } from '@/api/mutationFeedback';
+import { runWithMinimumPendingTime } from '@/api/utils/mutationFeedback';
 import { taxAdvantagedCategoryKeys } from '@/api/queryKeys';
 import {
   refreshTaxAdvantagedCategoryLimitCaches,
